chore: remove old CRUD module
